implement backpressure

+++ src/main/java/com/github/davidmoten/rtree/NonLeaf.java
@@ -76,7 +76,7 @@ final class NonLeaf<T> implements Node<T> {
     @Override
     public ImmutableStack<NodePosition<T>> search(Func1<? super Geometry, Boolean> condition,
             Subscriber<? super Entry<T>> subscriber, ImmutableStack<NodePosition<T>> stack,
-            int request) {
+            long request) {
         Preconditions.checkArgument(!stack.isEmpty());
         NodePosition<T> np = stack.peek();
         Preconditions.checkArgument(this == np.node());
